formularios y cambios en la db

+++ database/migrations/2022_02_21_201115_create_tareas_table.php
@@ -16,7 +16,7 @@ class CreateTareasTable extends Migration
         Schema::create('tareas', function (Blueprint $table) {
             $table->id();
             $table->text('descripcion');
-            $table->integer('elementos');
+            $table->float('elementos');
 
             $table->unsignedBigInteger('id_user')->nullable();
             $table->foreign('id_user')
